requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
enum34==1.1.6
future==0.16.0
idna==2.5
ipaddress==1.0.18
oauthlib==2.0.2
pycparser==2.17
pyOpenSSL==17.0.0
pysolr==3.6.0
python-twitter==3.3
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
watson-developer-cloud==0.26.0
